updates according to PR review
