+++ common/mount.py
@@ -143,7 +143,7 @@ class Mount(object):
                  profile_id = None,
                  tmp_mount = False,
                  parent = None):
-        self.config = cfg if cfg is not None else config.Config()
+        self.config = cfg or config.Config()
         self.profile_id = profile_id if profile_id is not None else self.config.currentProfile()
         self.tmp_mount = tmp_mount
         self.parent = parent
@@ -431,7 +431,7 @@ class MountControl(object):
         self.mountproc = None
         self.symlink_subfolder = None
 
-        self.config = cfg if cfg is not None else config.Config()
+        self.config = cfg or config.Config()
 
         self.profile_id = profile_id if profile_id is not None else self.config.currentProfile()
