+++ qt/qttools.py
@@ -139,6 +139,21 @@ def set_wrapped_tooltip(widget: QWidget,
     widget.setToolTip('\n'.join(result))
 
 
+
+def update_combo_profiles(config, combo_profiles, current_profile_id):
+    """
+    Updates the combo box with profiles.
+
+    :param config: Configuration object with access to profile data.
+    :param combo_profiles: The combo box widget to be updated.
+    :param current_profile_id: The ID of the current profile to be selected.
+    """
+    profiles = config.profilesSortedByName()
+    for profile_id in profiles:
+        combo_profiles.addProfileID(profile_id)
+        if profile_id == current_profile_id:
+            combo_profiles.setCurrentProfileID(profile_id)
+
 # |---------------------|
 # | Misc / Uncatgorized |
 # |---------------------|
@@ -703,3 +718,4 @@ class ProfileCombo(SortedComboBox):
             if self.itemData(i) == profileID:
                 self.setCurrentIndex(i)
                 break
+
